// frontend/src/app/generated/api/model/languageProficiency.ts
/**
 * Yaam.API | v1
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export const LanguageProficiency = {
  Basic: 'Basic',
  BusinessProficiency: 'BusinessProficiency',
  Fluent: 'Fluent',
  Native: 'Native',
} as const;
export type LanguageProficiency = (typeof LanguageProficiency)[keyof typeof LanguageProficiency];